events card revamp
