+++ pubspec.yaml
@@ -35,6 +35,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   toggle_switch: ^2.0.1
+  flutter_rating_bar: ^4.0.1
   #google_fonts: ^3.0.1
 
 dev_dependencies:
